requirements.txt
asttokens==2.0.5
backcall==0.2.0
click==8.1.3
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
Flask==2.1.2
fonttools==4.33.3
imageio==2.19.3
importlib-metadata==4.11.4
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
networkx==2.8.4
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
stack-data==0.2.0
tifffile==2022.5.4
torch==1.11.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
